Fixes #22594: Allow listing upstream subs
Refactor upstream candlepin resources
