+++ app/views/katello/api/v2/upstream_subscriptions/base.json.rabl
@@ -1,0 +1,12 @@
+object @resource ||= @object
+
+attributes :pool_id
+attributes :status
+attributes :quantity
+attributes :start_date
+attributes :end_date
+attributes :contract_number
+attributes :consumed
+attributes :product_name
+attributes :product_id
+attributes :subscription_id
