+++ config/routes/api/v2.rb
@@ -317,6 +317,8 @@ Katello::Engine.routes.draw do
               put :refresh_manifest
             end
           end
+
+          api_resources :upstream_subscriptions, only: :index
         end
 
         api_resources :host_collections
