+++ test/lib/tasks/update_subscription_facet_backend_data_test.rb
@@ -16,6 +16,8 @@ module Katello
       Katello::Host::SubscriptionFacet.where("id != #{@host.subscription_facet.id}").destroy_all
       Katello::Host::SubscriptionFacet.any_instance.stubs(:host).returns(@host)
       Katello::Candlepin::Consumer.any_instance.stubs(:consumer_attributes).returns(:facts => {:foo => :bar})
+      Katello::Resources::Candlepin::Consumer.stubs(:virtual_guests).returns([])
+      Katello::Resources::Candlepin::Consumer.stubs(:virtual_host).returns(nil)
 
       Katello::Host::SubscriptionFacet.expects(:update_facts).with(@host, :foo => :bar).once
